Type: fix 19453
Description: Occasional delete dialog residue.

RootCause: After clicking on the confirmation in the delete confirmation
dialog, execute the function of close, and then send a signal that pops
up the delete progress dialog box. Because the slot is connected by the
default Qt:AutoConnection, the actual execution logic is to pop up the
delete progress dialog and close the delete confirmation dialog.

Solution: Use QTimer to send the singal of popUninstallDialog.