Type: fix 19657
Description: After starting the software with the sudo command, the
system fonts are deleted.

RootCause: After starting the software with the sudo command,
QDir::homePath is no longer the user directory. The database does not
exist, causing the data to be out of sync. Delete the system fonts.

Solution: When the user is root, do not delete system fonts.