Type: fix 20011
Description: Text line spacing is to large in deleteing dialog. And, the
background color of deleteting dialog is do not match UI.
Refactor the code of deleteing dialog.

RootCause: The size of deleteing dialog is not fixed. Background color is
not set.

Solution: Fix dialog size and set the color of background.